+++ service/controller/resource/nodeautorepair/create.go
@@ -11,7 +11,6 @@ import (
 	"github.com/giantswarm/microerror"
 	corev1 "k8s.io/api/core/v1"
 
-	"github.com/giantswarm/aws-operator/pkg/project"
 	"github.com/giantswarm/aws-operator/service/controller/controllercontext"
 	"github.com/giantswarm/aws-operator/service/controller/key"
 )
@@ -52,7 +51,7 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 			K8sClient: cc.Client.TenantCluster.K8s.CtrlClient(),
 			Logger:    r.logger,
 
-			LockName:              project.Name(),
+			LockName:              Name,
 			NotReadyTickThreshold: nodeTerminationTickThreshold,
 		}