+++ pkg/annotation/annotation.go
@@ -4,5 +4,5 @@ const (
 	Docs                    = "giantswarm.io/docs"
 	InstanceID              = "aws-operator.giantswarm.io/instance"
 	MachineDeploymentSubnet = "machine-deployment.giantswarm.io/subnet"
-	AWSMetadata             = "beta.giantswarm.io/aws-metadata-v2"
+	AWSMetadata             = "alpha.giantswarm.io/aws-metadata-v2"
 )