+++ service/controller/resource/tccpn/template/template_main_launch_template.go
@@ -31,6 +31,8 @@ const TemplateMainLaunchTemplate = `
           Name: !Ref ControlPlaneNodesInstanceProfile
         ImageId: {{ .Instance.Image }}
         InstanceType: {{ .Instance.Type }}
+        MetadataOptions:
+          HttpTokens: {{ .Metadata.HttpTokens }}
         Monitoring:
           Enabled: {{ .Instance.Monitoring }}
         NetworkInterfaces: