+++ flag/service/cluster/dockerhub/dockerhub.go
@@ -1,0 +1,5 @@
+package dockerhub
+
+type Dockerhub struct {
+	Token string
+}