+++ service/controller/control_plane.go
@@ -64,6 +64,7 @@ type ControlPlaneConfig struct {
 	ClusterDomain             string
 	ClusterIPRange            string
 	DockerDaemonCIDR          string
+	DockerhubToken            string
 	ExternalSNAT              bool
 	HostAWSConfig             aws.Config
 	IgnitionPath              string
@@ -221,6 +222,7 @@ func newControlPlaneResources(config ControlPlaneConfig) ([]resource.Interface, 
 				ClusterDomain:             config.ClusterDomain,
 				ClusterIPRange:            config.ClusterIPRange,
 				DockerDaemonCIDR:          config.DockerDaemonCIDR,
+				DockerhubToken:            config.DockerhubToken,
 				ExternalSNAT:              config.ExternalSNAT,
 				IgnitionPath:              config.IgnitionPath,
 				ImagePullProgressDeadline: config.ImagePullProgressDeadline,