+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Annotation `alpha.giantswarm.io/aws-metadata-v2` to enable AWS Metadata API v2
+- Annotation `alpha.giantswarm.io/aws-subnet-size` to customize subnet size of Control Plane and Node Pools
 
 ### Fixed